style: add clang-format support
Ensuring a consistent style is important for readability and
maintainability. This project follows the Google C++ Style Guide and
therefore requires a mechanism to enforce this style across different
platforms and environments.

To address this, clang-format is used which is a well-known cross
platform formatting tool with support for the Google C++ Style Guide.

Note that automated formatting is limited to tools which support
.clang-format files (e.g. IDEs). Ideally, this would be run prior to
check-in via Git commit hooks which is currently tracked by #9.

Closes #3