test: enable running tests using CTest
Currently, unit tests are run from a separate executable which is built
with the project. Although this is sufficient, CMake provides support
for running unit tests using CTest which has additional benefits such as
CI integration, IDE support, and configuration using CMakePresets.

Catch2 provides support for for CTest through CMake scripts that help
with automatic test registration. As a result, integration with CTest is
as simple as including those scripts and calling catch_discover_tests
with the target test executable.

Closes #8