style: use two character indent width
One of the strengths and weakness of Vulkan is that it is excessively
verbose which can make code difficult to read with excessive line
wrapping.

This change reduces the indent width from four to two
characters which allows for more code to fit on a single line while
retaining readability. This more closely aligns the project coding style
with the Google C++ Style Guide by removing the four space indent width
exception.