build: disable global target compile options
Currently, compiler options to enable all warnings as errors and address
sanitizer in debug builds are added globally for all targets in the
project. This is considered bad practice because it forces any future
dependencies to also compile with those compiler options which is can
create problems for untested build toolchains.

To fix this, CMake scripts were updated to replace add_compile_options
and add_link_options with target_compile_options and target_link_options
respectively using the PRIVATE keyword to prevent propagating target
properties to downstream dependencies. This unfortunately creates a new
problem where compiler options previously added to all targets must be
repeated on a per target basis which is both error-prone and tedious. To
mitigate this, compiler and link options are reused by creating an
INTERFACE library target where relevant options are added which can then
be linked to on a per target basis.

This commit also takes the opportunity to refactor existing CMake
scripts to be more concise and remove redundancies.

This stack overflow post inspired the use of an interface library target
to reuse common compiler and link options:
https://stackoverflow.com/questions/60041896/reuse-target-compile-options-from-variable-for-multiple-targets-cmake