feat: select a Vulkan physical device
To create a logical device in Vulkan, a physical device must first be
selected.

Although there is a number of criteria to select a physical device from,
this implementation opts for the simplest reliable approach of using a
physical device with graphics and presentation capabilities that favors
discrete GPUs.

Relates to #25