feat: create a Vulkan swapchain
To present images to a window surface, a Vulkan swapchain must first be
created which acts as a queue for presenting rendered images.

Creating a swapchain depends on a number of configurations at the
intersection of physical device and surface capabilities such as
presentation mode, image format, image extent, and image count. Upon
completion, swapchain image views are created for later use in the
graphics pipeline.

Documentation for querying for surface capabilities can be found here:
https://docs.vulkan.org/spec/latest/chapters/VK_KHR_surface/wsi.html#_surface_queries