feat: create a Vulkan shader module
One of the first steps in creating a graphics pipeline in Vulkan is to
create Vulkan shader modules which represent thin wrappers around SPIR-V
generated shaders.

To compile shaders to SPIR-V, a dependency on glslang is added to
compile GLSL shaders at runtime. Unfortunately glslang isn't well
documented and a runtime implementation is primarily adapted from the  C
functional interface listed on their GitHub's README. Additional design
considerations include:
- Using a singleton to ensure glslang is only initialized once
- Specializing std::formatter<glslang_stage_t> to improve error messages
- Using streaming operators to print shader/program info logs
The current implementation does not attempt to generate shader debug
symbols although it was explored briefly and is tracked by #34 for
further consideration/testing once a functional graphics pipeline is
working.

One complication encountered during development was that default shaders
which are stored in the engine source tree need to be copied to a game's
build tree which complicates loading shaders with relative file paths
because they are located in different directories in the final build
output. This was mitigated adding a custom CMake target to copy default
shaders from the engine source tree to a game's build tree. This is
admittedly suboptimal because it requires (a) hardcoding filepaths in
CMakeList.txt files and (b) duplicating for each new game sample.
Developing a robust solution for this is outside the scope of this PR
and is currently tracked by #37.

Shader compilation was confirmed to work on x64 Debug/Release and x86
Release. It unfortunately does not work on x86 Debug builds due to an
ASan violation in glslang. Because this does not reproduce in x64 Debug
builds, it's assumed to be a defect in glslang although this requires
further investigation which is tracked by #41.

http://github.com/KhronosGroup/glslang#c-functional-interface-new

Relates to #34, #37, #41
Closes #35, #42