style: always name function parameters
Parameter names are not always used for function definitions located in
header files when the type is obvious (e.g., vk::Instance instance).
This introduces inconsistency with function definitions which must use
named parameter names due to ambiguity (e.g., int width, int height).
This is also required by the clang-tidy readability-named-parameter
check.

The decision on when to use parameter names based on subjective criteria
of what is obvious or not creates inconsistency. Therefore, the decision
on parameter names moving forward is to always use named parameters in
function definitions.

Closes #40