docs: add GitHub issue templates
Creating issues in GitHub lacks any structure which impacts the ability
to fully evaluate all the implications of a bug or feature.

To address this, two issue templates are added: one for bug reports and
another for general work tracking. A dedicated feature request template
was not included because the project is not yet public, but should be
considered if/when that happens.

Documentation for how to create a custom GitHub issue template can be
found here:
https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository

Closes #39