build: do not link against vulkan-1.lib
Vulkan-Hpp provides a dynamic loader implementation which removes the
need to manually link against vulkan-1.lib. Thefore, this can be safely
removed from CMakeList.txt.

Furthermore, this change addresses outdated or incorrect information in
the project's README (including use of the a dynamic loader). As part of
this initiative, the .gitmessage commit template has been reinstated.
This was previously removed because it cannot be integrated in an
automated fashion. However, after further consideration, the benefits
having using such a template outweight the need to manually configure
during initial project setup.

This was tested and confirmed to work in release builds without the
Vulkan SDK installed on Windows x64. Debug builds were also verified
after reinstalling the Vulkan SDK.