build: add pre-commit hooks
When committing to a branch, there is no enforcement of project coding
standards. Pre-commit hooks are an official solution to run arbitrary
scripts to enforce those standards but must be added to the .git/hooks
directory which is unfortunately not managed by version control.

pre-commit is a python package that solves this problem by enabling
sharing of common git commit hooks. To use simply run pre-commit install
 which will download and install hooks from git repositories listed in
 .pre-commit-config.yaml which will run the next time git commit is run.

This commit adds an initial set of pre-commit hooks. While some are
small in scope (e.g., trailing whitespace), the two most important are
no-commit-to-branch which prevents checking in directly to main and
clang-format which will ensure a consistent code style. This also
enables future support for clang-tidy, cpplint, and IWYU which have been
deferred to keep this change focused.

More documentation on pre-commit has be found here:
https://pre-commit.com/index.html

Relevant C++ pre-commits can be found here:
https://github.com/pocc/pre-commit-hooks

Closes #9
Relates to #3, #4