style: add cpplint pre-commit hook
This project follow the Google C++ Style Guide. While many of its
guidelines can be enforced with clang-format, there are some which must
be enforced with a static analyzer.

cpplint is a python based tool developed by Google to ensure C++ code
conforms to the Google Style Guide. This is being as a pre-commit hook
to avoid increasing build times and avoid an explicit dependency on
python for users who only interested in building, but not contributing
to the project.

Documentation on the cpplint pre-commit hook can be found here:
https://github.com/cpplint/cpplint

Related to #9.