docs: update commit message format
The current commit message template, while thorough, requires excessive
information that would be more appropriate for a PR description. This
unnecessarily encumbers the process of committing changes thereby
slowing down progress.

This change streamlines this process by changing the commit message
format to require three sections: a header, body, and optional footer.
The header remains the same, while the body allows as little as one
paragraph explaining the motivation behind the change and how the commit
addresses it. Lastly, the footer remains the same as a place to link to
GitHub issues. This format is enforced with an updated commit message
template in addition to updating the CONTRIBUTING guidelines.

This change was largely inspired by Angular's commit message format:
https://github.com/angular/angular/blob/main/.gitmessage

Relates to #6