feat: add a data view abstraction
Buffer<T>::Copy only supports lists which is problematic for situations
where only a single value needs to be copied (e.g., uniform buffer).
One solution to this problem is to maintain two separate, but identical,
APIs that handle copying single values versus a list. However, a better
approach is to create an abstraction that can handle both a single item
and a list. Data view is such an abstraction that works like std::span
but can also handle single values. It achieves this in a scalable way by
using range concepts to accept any type of contiguous range in addition
adding overloads for a single item. It's worth noting that
vk::ArrayProxy behaves in a similar way but does not have support for
determining the size in bytes.