refactor: use vk::ArrayProxy to copy buffer data
Buffer<T>::Copy only supports lists which is problematic for situations
where only a single value needs to be copied (e.g., uniform buffer).
One solution to this problem is to maintain two separate, but identical,
APIs that handle copying single values versus a list. However, a better
approach is to create an abstraction that can handle both a single item
and a list. vk::ArrayProxy is such an abstraction that works like
std::span but can also handle single values.