feat: enable depth testing
When rendering overlapping polygons, the order in which they are drawn
directly influences their visiblity regardless if one is farther away
than another. One approach to this problem is to draw polygons from back
to front. However, the more standard solution is to use depth testing
which uses a separate buffer to keep track of the last known z-position
rendered to a given pixel and only update it if a closer z-position is
identified when rendering multiple meshes.

In Vulkan, the depth buffer is represented by a VkImage that is bound to
device memory (similar to a VkBuffer). Once created, the depth testing
is enabled through referencing the depth buffer in a render pass and
included as a framebuffer depth attachment.