feat: add support for loading .obj models
After rendering a simple triangle, the next logical step is to render a
3D mesh which is now possible after enabling depth testing. Wavefront
.obj is a simple file format that can be used to describe a mesh with
vertex positions, texture coordiantes, and normals. This commit adds
initial support for loading models using a subset of the file format
specification. Because material support is not yet available meshes are
rendered using interpolated normals for their end fragment color.