feat: update model transform using push constants
Push constants provide an efficient mechanism for sending a small amount
of data with command buffers which are ideal for per-frame, per-model
data such as model transforms. To achieve this, gfx::Mesh is refactored
to include a 4x4 model transform with corresponding APIs to translate,
rotate, and scale the mesh which is then applied to command buffer when
rendering the mesh.