feat: add mesh translation, rotation, and scale
A compulsory feature of any model viewer is to be able to translate,
rotate, and scale a mesh. To achieve this, window event handlers are
created to enable rotating and translating a mesh using the left and
right mouse buttons respectively in addition to uniformly scaling the
mesh when scrolling.