refactor: move game logic out of scene class
There are several flaws in the design of gfx::Scene which are admittedly
due to the small scope of this project. First, it includes "game" logic
(i.e., keyboard and mouse event handlers) that would ideally be reserved
for a separate class. Second, it doesn't own a gfx::Window yet requires
lambda capturing a pointer to it which may outlive its lifetime.

This commit refactors improves upon the existing design by introducing a
new Game abstraction which is used to set up the scene and game logic.
Ideally, the scene abstraction would remain, but the limited scope of
the project makes it hard to justify. However, this may be revisted as
the project complexity grows to include lighting and materials.