feat: add a data view abstraction
vk::ArrayProxy is a useful abstraction because, unlike std::span, it
provides the ability to represent both a contiguous range of data or
a single element. Despite this, it is unwise to rely on a third party
utility as part of a public API because it is subject to change and can
easily be disabled by defining VULKAN_HPP_DISABLE_ENHANCED_MODE.

This commit introduces a similar abstraction named data view that
improves upon vk::ArrayView by utilizing C++20 range concepts and adding
an API to calculate its total size in bytes (similar to std::span).