style: use vec3 for mesh translation/scale APIs
Mesh translation and scale APIs use individual x/y/z parameters which
are then converted to vec3 so they can be used with glm matrix transform
APIs. This results in an awkward usage when x/y/z parameters are already
in vec3 form (e.g., the existing game logic for handling cursor position
changes). It also requires repeating the same value three times to
achieve uniform scale which is exclusively used in the project.

These API deficiencies are all resolved by simply using vec3 as part of
the mesh translation and scale APIs which makes the code easier to read
and more natural to use.