feat: add lighting using phong reflection
Phong reflection is a standard lighting model that approximates light
distribution by accounting for the interaction of ambient, diffuse,
specular, and reflectance properties of a material with the light,
viewing, and normal directions. For now, these properties are hardcoded
as constants in the fragment shader but will be converted into
appropriate uniform buffers or push constants in a subsequent commit.