feat: add mesh simplification implementation
This commit adds an implementation of Surface Simplification using
Quadric Error Metrics by Garland-Heckbert.

The core idea behind the algorithm is to use a process known as edge
contraction to collapse an edge into a single vertex that maximally
preserves the shape of the mesh. This vertex position can be solved for
analytically by minimizing the squared distance of adjacent triangles
that are attached to the new vertex. Using squared distance as a cost
metric, edges can then be sorted using a priority queue to process edges
by min cost until the mesh has been sufficiently simplified.

This algorithm has a few limitations:
  - It only works on watertight meshes
  - It does not handle updates to texture coordinates.

For a more detailed description of the algorithm see the PDF located at
docs/surface_simplification.pdf