build: create geometry and graphics libraries
Despite the app being divided into logical subdirectories, it is still
built as a single executable with no shared libraries which precludes
testing without recompiling each translation unit.

This commit addresses this by creating static libraries for geometry and
graphics subdirectories which are linked against the main application
executable. It also takes the liberty to create interface targets to
privately link against interface libraries which enable all warnings as
errors and enable ASan in debug builds which is considered a CMake best
practice.