test: add geometry and graphics unit tests
This commit adds an initial set of unit tests for the geometry and
graphics libraries. Note that some tests depend on instantiating a
graphics device to load a mesh for testing which is unfortunately
required to test APIs without introducing awkward semantics and
attenuating class invariants. Lastly, including of some .cpp files is
intentional to test methods internal to that translation unit.