feat: add arc camera implementation
This commit adds initial support for an arc camera which orbits around
a given target. To simplify the implementation, the camera position is
specified in spherical coordinates which is converted to cartesian
coordinates when constructing the view transform.

Note that this achieves the same effect as the arcball implementation
but uses a different perspective of moving the camera around the mesh
rather than rotating it using an axis-angle constructed from a unit
sphere. This approach simplifies the implementation because it does not
require constructing the inverse of the model-view transform to apply
the arcball rotation to the mesh. To avoid duplicate logic, the arcball
implementation is therefore being deprecated.