feat: enable arc camera zoom using scroll input
One critical piece of functionality missing from the arc camera
implementation is the ability to zoom on its specified target. This
functionality is commonly found in other implementations and is
especially important for viewing fine details for model viewers.

This commit addresses this by adding a new zoom API to the arc camera
which is achieved by modifying its spherical position's radius by a
rate determined using scroll input.