refactor: simplify device creation
One of the primary limitations with the current design for physical
device selection and subsequent logical device creation is leaking
internal implementation details (i.e, RankedPhysicalDevice) which was
primarily due to not having a default constructor for the queue family
abstraction (explicitly omitted to preserve class invariants).

This commit reworks the design to avoid leaking this implementation
details by creating a physical device abstraction during the ranking
process rather than afteward. This commit also removes abstractions for
the queue and queue family in favor of a more simple wrapper around
queue family indices which significantly simplifies the overall design.