refactor: add an explicit subpass dependency
Currently, submission to the graphics queue will wait on the top of pipe
pipeline stage flag. This is suboptimal because it can instead wait on
the color attachment output stage (in conjunction with the creation of
an explicit subpass dependency) which enables further execution up to
that stage which can improve performance. This is also recommended as a
best practice on the Vulkan-Samples repository.