refactor: don't store transient attachments
A recommended best practice in Vulkan is to avoid invoking store
operations for transient render pass attachments which only live for the
duration of the frame. While the multisample color attachment correctly
avoid this, the depth attachment does not. This commit addresses this by
changing the store operation for the depth attachment from eStore to
eDontCare.

This commit also adds the eTransient image usage flag bit to the
multisample color and depth attachments which is also a recommended best
practice when used with the eLazilyAllocated memory property flag bit.
Unfortunately, this memory flag bit only appears to be available on
mobile devices so it's unclear if this will affect any behavior change.
Regardless, it is included because it is both self documenting and can't
hurt the driver to have more information on its usage.

See https://github.com/KhronosGroup/Vulkan-Samples/blob/main/samples/performance/msaa/README.adoc
for more details.