build: move headers into include directories
This commit reworks the project source tree to move headers into
dedicated include directories which is considered a best practice in
CMake to avoid the possibility of directly including source cpp files.
It also removes the math library in favor of consolidating with the
graphics library due to it only containing a single file.