style: wrap source definitions in a namespace
The Google C++ Style Guide states that "namespaces wrap the entire
source file after includes." This is given in the broader context of
requiring namespaces and although a specific justification isn't given,
it's assumed to be recommended because it reduces typing over adding
individual namespace qualifiers to each function definition in source
files.

This commit updates the project code style to use this convention. Note
that use of anonymous namespaces in source files is also retained.