fix: prevent potential dangling pointer
When the graphics and present queue family indices differ, it's possible
for the Vulkan implementation to dereference a dangling pointer due to
creating a pointer to a stack allocated array which goes out of scope
before it is used when creating a swapchain. This was also technically
possible when the graphics and queue family indices were the same, but
it turns out that pointer in question was never used in that situation
per the Vulkan specification.