build: move headers next to source files
There are a few different approaches to managing header files in CMake.
One is to segregate header files into a separate include directory while
another encourages header and source files to be places in the same
directory. Until now, this project has taken the former approach and
while it has the advantage of providing a more organized file structure
it can also make it more difficult to navigate and maintain the project
by avoiding the need to jump between deeply nested directories to edit
closely related files.