build: manage header files using target_sources
Header files are not required to be explicitly listed in CMakeLists.txt
files because they are transitively added in source files that include
them. While sufficient, this approach has the disadvantage of providing
a suboptimal tooling experience for IDEs such as Visual Studio which do
not list header files when viewing from the CMake targets view.

To address this, header files are now included in CMakeLists.txt files
using the newer FILE_SET mechanism in conjunction with target_sources.
This has the additional benefit of enabling a more fine grained control
of headers over target_include_sources when headers and source files
reside in the same directory (recently introduced in commit 7a5aaa8c).