refactor: rework glslang compiler implementation
The current glslang implementation requires using the singleton pattern
in its API which is an awkward user experience. This commit reworks the
glslang compiler implementation to use a single free function as part of
its public interface. It also provides various improvements to the
implementation including simplifying logging during various steps of the
compilation process.