build: create an assets symlink
When compiling a project, the working directory the game executable runs
(i.e., the cmake binary directory) is different from the directory where
source files live (i.e., the cmake source directory). This creates a
problem for files that are not compiled with the project (e.g., models,
shaders) because they not part of the generated build directory. The
current approach to solve this is to simply copy the entire assets
directory from the source to binary directories. While sufficient, it's
inefficient for the purposes of developing. Instead a more lightweight
approach is taken by creating a symlink using the cross-platform cmake
command. While there is an argument to be made that the prior approach
is better for distributing an executable, there are better tools to
solve that and it's outside the scope of this project anyway.